pyproject.toml
[project]
name = "cloudnimbus"
version = "1.0.0"
description = "Aerial roof measurement and auto-estimating platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "pydantic==2.9.0",
    "openai==1.50.0",
    "httpx==0.27.0",
    "Pillow==10.4.0",
    "shapely==2.0.6",
    "pyproj==3.7.0",
    "python-dotenv==1.0.1",
    "numpy==1.26.4",
    "torch==2.4.0",
    "torchvision==0.19.0",
    "python-multipart==0.0.9",
    "jinja2==3.1.4",
    "weasyprint==62.3",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "streamlit>=1.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

# --- Ruff ---
[tool.ruff]
target-version = "py312"
line-length = 120
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "N"]
ignore = [
    "E501",  # line length handled by formatter
    "B008",  # FastAPI Depends() pattern
    "N805",  # Pydantic validator cls
]

[tool.ruff.lint.isort]
known-first-party = ["config", "models", "pipeline"]

# --- mypy ---
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "shapely.*",
    "pyproj.*",
    "PIL.*",
    "google.*",
    "weasyprint.*",
]
ignore_missing_imports = true
